Fixing circular dependency between common and legacy
