+++ src/neptune/common/envs.py
@@ -18,3 +18,11 @@
 API_TOKEN_ENV_NAME = "NEPTUNE_API_TOKEN"
 
 NEPTUNE_RETRIES_TIMEOUT_ENV = "NEPTUNE_RETRIES_TIMEOUT"
+
+PROJECT_ENV_NAME = "NEPTUNE_PROJECT"
+
+NOTEBOOK_ID_ENV_NAME = "NEPTUNE_NOTEBOOK_ID"
+
+NOTEBOOK_PATH_ENV_NAME = "NEPTUNE_NOTEBOOK_PATH"
+
+BACKEND = "NEPTUNE_BACKEND"
